backend/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.37.0
python-multipart==0.0.20
PyMuPDF==1.25.5
openai==1.107.0
chromadb==1.0.20
python-dotenv==1.1.1
